fapiao2/pyproject.toml
[project]
name = "fapiao2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "pyinstaller>=6.17.0",
    "python-dotenv>=1.2.1",
    "selenium>=4.39.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
]
